Move ResourceInfo to data_structures subpackage
