+++ src/benchq/data_structures/__init__.py
@@ -14,3 +14,18 @@ from .hardware_architecture_models import (
     BasicArchitectureModel,
 )
 from .quantum_program import QuantumProgram, get_program_from_circuit
+from .resource_info import ResourceInfo
+
+__all__ = [
+    "AlgorithmImplementation",
+    "get_algorithm_description_from_circuit",
+    "DecoderModel",
+    "ErrorBudget",
+    "GraphPartition",
+    "BASIC_ION_TRAP_ARCHITECTURE_MODEL",
+    "BASIC_SC_ARCHITECTURE_MODEL",
+    "BasicArchitectureModel",
+    "QuantumProgram",
+    "get_program_from_circuit",
+    "ResourceInfo",
+]
